% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AgeTopicModels-package.R
\docType{package}
\name{AgeTopicModels-package}
\alias{AgeTopicModels}
\alias{AgeTopicModels-package}
\title{AgeTopicModels: Inferring Age-Dependent Disease Topic from Diagnosis Data}
\description{
We propose an age-dependent topic modelling (ATM) model, providing a low-rank representation of longitudinal records of hundreds of distinct diseases in large electronic health record data sets. The model assigns to each individual topic weights for several disease topics; each disease topic reflects a set of diseases that tend to co-occur as a function of age, quantified by age-dependent topic loadings for each disease. The model assumes that for each disease diagnosis, a topic is sampled based on the individual’s topic weights (which sum to 1 across topics, for a given individual), and a disease is sampled based on the individual’s age and the age-dependent topic loadings (which sum to 1 across diseases, for a given topic at a given age). The model generalises the Latent Dirichlet Allocation (LDA) model by allowing topic loadings for each topic to vary with age. Jiang (2023) \doi{https://doi.org/10.1038/s41588-023-01522-8}
}
\author{
\strong{Maintainer}: Xilin Jiang \email{xilinjiang@hsph.harvard.edu} (\href{https://orcid.org/0000-0001-6773-9182}{ORCID})

}
\keyword{internal}
