\name{ATM-package}
\alias{ATM-package}
\alias{ATM}
\docType{package}
\title{
\packageTitle{ATM}
}
\description{
\packageDescription{ATM}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{ATM}
\packageIndices{ATM}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{ATM}

Maintainer: \packageMaintainer{ATM}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
