\name{atm}
\alias{atm}
\alias{atmx}
\alias{evaluate.atm}
\alias{fit.atm}
\alias{gf.atm}
\alias{oospfitted.atm}
\alias{oospprint.summary.atm}
\alias{oospresiduals.atm}
\alias{oospsummary.atm}
\title{Additive Term Models}
\description{Create atm objects.}
\usage{
atm (y, \dots, fit=TRUE, rs=TRUE, maxfits=25, w=NULL)
atmx (m)
\method{evaluate}{atm}(m, us, \dots)
\method{fit}{atm}(m, rs=TRUE, maxfits=25, intermediate.diagnostics=FALSE, \dots)
\method{gf}{atm}(m, \dots)
\method{oospfitted}{atm}(m, \dots)
\method{oospprint}{summary.atm}(s, \dots)
\method{oospresiduals}{atm}(m, y=m$y, partial=NULL, \dots)
\method{oospsummary}{atm}(m, which=NA, \dots)
}
\arguments{
\item{fit}{}
\item{rs}{}
\item{maxfits}{}
\item{w}{}
\item{m}{}
\item{us}{}
\item{intermediate.diagnostics}{}
\item{s}{}
\item{y}{}
\item{partial}{}
\item{which}{}
\item{\dots}{}
}


