% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta.R
\name{C2K}
\alias{C2K}
\title{Convert degrees C to Kelvin}
\usage{
C2K(C)
}
\arguments{
\item{C}{numeric value of temperature in degrees C.}
}
\value{
\code{C2K} returns the equivalent temperature in Kelvin.
}
\description{
Convert degrees C to Kelvin
}
\author{
Dan Kelley
}
