% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.R
\name{cal_extract}
\alias{cal_extract}
\title{Extract calibration results.}
\usage{
cal_extract(filename, vessel.name, survey.name, cal.group = "SWFSC ")
}
\arguments{
\item{filename}{An string containing the name of the calibration file.}

\item{vessel.name}{A string containing the vessel name.}

\item{survey.name}{A string containing the surven name.}

\item{cal.group}{A string containing the name of the calibration group.}
}
\value{
A list containing calibration information, results, and ping data.
}
\description{
Extract calibration results.
}
