% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{calc_dist}
\alias{calc_dist}
\title{Calculate distance to a feature}
\usage{
calc_dist(x1, y1, x2, y2)
}
\arguments{
\item{x1}{Origin longitude.}

\item{y1}{Origin latitude.}

\item{x2}{Destination longitude.}

\item{y2}{Destination latitude.}
}
\value{
A vector containing the distance from origin to destination.
}
\description{
Calculate distance to a feature
}
\examples{
calc_dist(long1, lat1, long2, lat2)
}
