% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological.R
\name{calculate_age}
\alias{calculate_age}
\title{Estimate Pacific sardine age}
\usage{
calculate_age(standardLength_mm)
}
\arguments{
\item{standardLength_mm}{A vector containing the species' standard length (mm).}
}
\value{
A vector of age groups for age-0 (0) and age-1+ (1).
}
\description{
Estimates the age of Pacific sardine (Sardinops sagax) from their standard length (SL, mm)
}
\examples{
calculate_age(standardLength_mm, 150)
}
