% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{calculate_depth_integrated_average}
\alias{calculate_depth_integrated_average}
\title{Calculate depth-integrated average values}
\usage{
calculate_depth_integrated_average(df, value_col, depth_col = "Depth")
}
\arguments{
\item{df}{Data frame containing data.}

\item{value_col}{Input value column.}

\item{depth_col}{Depth column (default = "Depth").}
}
\value{
Depth-integrated average value(s).
}
\description{
Calculate depth-integrated average values
}
