% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{calculate_fg_absorption}
\alias{calculate_fg_absorption}
\title{Calculate FG absorption}
\usage{
calculate_fg_absorption(T, S, D, f)
}
\arguments{
\item{T}{Temperature value(s).}

\item{S}{Salinity value(s).}

\item{D}{Density value(s).}

\item{f}{Some variable I don't know what it is.}
}
\value{
A value for absorption (alpha).
}
\description{
Calculate FG absorption
}
