% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{clone_estimATM}
\alias{clone_estimATM}
\title{Clone estimATM repository}
\usage{
clone_estimATM(
  repo_url = "https://github.com/kstierhoff/estimATM",
  path,
  overwrite = FALSE
)
}
\arguments{
\item{repo_url}{URL for estimATM repository on GitHub.}

\item{path}{Path to desired project location.}

\item{overwrite}{Overwrite existing directory.}
}
\value{

}
\description{
Clone estimATM repository
}
