% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATM_inference.R
\name{diseasematrix2longdata}
\alias{diseasematrix2longdata}
\title{Disease matrix reformatting for ATM}
\usage{
diseasematrix2longdata(disease_matrix)
}
\arguments{
\item{disease_matrix}{a disease matrix with the first column name "eid", other column are disease names. Disease should be coded as 0,1.}
}
\value{
a data frame which can be feed into ATM_wrapper
}
\description{
Disease matrix reformatting for ATM
}
\examples{
disease_matrix <- longdata2diseasematrix(HES_age_example)
diseasematrix2longdata(disease_matrix)
}
