% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sounding_download.R
\name{downloadAtmosphericSounding}
\alias{downloadAtmosphericSounding}
\title{Download atmospheric sounding data}
\usage{
downloadAtmosphericSounding(
  date = Sys.Date() - 1,
  region = "naconf",
  stationName = "Shearwater"
)
}
\arguments{
\item{date}{A string for a date, in yyyy-mm-dd format, or some other form
that \code{\link[=as.POSIXlt]{as.POSIXlt()}} can convert to a date or datetime.
If not provided, the previous day is used (because an early-morning call
to this function might yield a failed download).}

\item{region}{a character value indicating the region. This defaults
to \verb{"naconf}.}

\item{stationName}{a character value indicating the station name, which
the present function converts to a station number by \code{\link[=stationNumber]{stationNumber()}}.}
}
\description{
Download atmospheric sounding data
}
\author{
Dan Kelley
}
