% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sounding_download.R
\name{downloadSounding}
\alias{downloadSounding}
\title{Download atmospheric sounding data}
\usage{
downloadSounding(
  date = Sys.Date() - 1,
  region = "naconf",
  stationName = "Shearwater"
)
}
\arguments{
\item{date}{A string for a date, in yyyy-mm-dd format, or some other form
that \code{\link[=as.POSIXlt]{as.POSIXlt()}} can convert to a date or datetime.
If not provided, the previous day is used (because an early-morning call
to this function might yield a failed download).}

\item{region}{a character value indicating the region. This defaults
to \code{"naconf"}.}

\item{stationName}{a character value indicating the station name, which
the present function converts to a station number by \code{\link[=stationNumber]{stationNumber()}}.}
}
\description{
The file is downloaded from Reference 1, by constructing a URL
that queries the server for a particular station and time. The
construction of that URL is based on reverse-engineering the
URLs that get created by clicking on a map.  This means
that \code{downloadSounding} is brittle, and will
fail if the server setup changes.  Unfortunately, this
guessing game seems to be the best we can do, since the website
does not provide information on the URL construction, let alone
an API.
}
\references{
\enumerate{
\item \url{https://weather.uwyo.edu/upperair/sounding.shtml}
}
}
\author{
Dan Kelley
}
