% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermodynamics.R
\name{dryLapseRate}
\alias{dryLapseRate}
\title{Dry lapse rate in degC / m}
\usage{
dryLapseRate(g = 9.7986, C_P = 1004)
}
\arguments{
\item{g}{acceleration due to gravity, with the default as listed in
Reference 2.}

\item{C_P}{numeric value of the specific heat of air at sea level,
with the default being as listed in Reference 2, converted to J/(kg*K).}
}
\value{
\code{dryLapseRate} returns the lapse rate, in degrees C (or Kelvin) per
metre.
}
\description{
This is the ratio of the acceleration due to gravity and the specific
heat of the air, according to (3.53) of Reference 2.
}
\references{
\enumerate{
\item Wallace, John M., and Peter Victor Hobbs. Atmospheric Science: An
Introductory Survey. 2nd ed. International Geophysics Series,
\item \url{https://pds-atmospheres.nmsu.edu/education_and_outreach/encyclopedia/adiabatic_lapse_rate.htm}
}
}
\author{
Dan Kelley
}
