% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermodynamics.R
\name{dTdp_moist}
\alias{dTdp_moist}
\title{dT/dp for moist adiabat}
\usage{
dTdp_moist(TK, p)
}
\arguments{
\item{TK}{temperature in K.}

\item{p}{pressure in Pa (100*mbar)}
}
\value{
\code{dTdp_moist} returns the derivative in K/Pa.
}
\description{
This is from pymeteo; I cannot find the equivalent formula at
in WH.
}
