% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermodynamics.R
\name{dTdz_moist}
\alias{dTdz_moist}
\title{dT/dz for moist adiabat}
\usage{
dTdz_moist(TK, p)
}
\arguments{
\item{TK}{temperature in K.}

\item{p}{pressure in Pa (100*mbar)}
}
\value{
\code{dTdz_moist} returns the derivative in K/m.
}
\description{
This is from pymeteo; I cannot find the equivalent formula at
in WH.
}
