% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/effects.R
\name{Effect.fitTMB}
\alias{Effect.fitTMB}
\title{Adapts package \code{effects}}
\usage{
\method{Effect}{fitTMB}(focal.predictors, mod, which_formula = "taxis", ...)
}
\description{
Calculate effects for plotting
}
