% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{estimate_inshore}
\alias{estimate_inshore}
\title{Extract acoustic intervals for nearshore extrapolation}
\usage{
estimate_inshore(nasc.df, feature.df, lat.radius = 0.1)
}
\arguments{
\item{nasc.df}{A data frame containing total backscatter from all CPS
(\code{cps.nasc}), acoustic proportions (\code{prop.spp}), and
weight-specific backscattering cross sections (\code{sigmawg.spp}) in a
given stratum}

\item{feature.df}{A data frame containing the latitude and longitude of the
feature used to reduce \code{nasc.df}}

\item{lat.radius}{A vector used to restrict the features being compared.}
}
\value{
A data frame containing backscatter equidistinant from a feature.
}
\description{
Extract acoustic intervals for nearshore extrapolation
}
\examples{
estimate_inshore(nasc.df, feature.df, lat.radius = 0.1)
}
