% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological.R
\name{estimate_length}
\alias{estimate_length}
\title{Estimate CPS length}
\usage{
estimate_length(scientificName, weightg, model.type = "OLS", season)
}
\arguments{
\item{scientificName}{A vector containing the species' scientific name (Clupea
pallasii, Engraulis mordax, Sardinops sagax, Scomber japonicus, or
Trachurus symmetricus).}

\item{weightg}{A vector containing specimen weight, in grams.}

\item{model.type}{The model used to estimate weight, either \code{OLS} (Ordinary Least Squares) or \code{GLM} (Generalized Linear Model).}

\item{season}{The season in which the specimen was collected, either \code{spring} or \code{summer}.}
}
\value{
Specimen total length (TL), in millimeters.
}
\description{
Calculate the estimated length of CPS species from a known weight measurement, using the mass-length relationship
models presented in Palance et al. 2019.
}
\examples{
estimate_length("Sardinops sagax", 5, model.type = "OLS", "summer")
}
