% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atm.R
\name{estimate_ts}
\alias{estimate_ts}
\title{Estimate target strength of CPS from total length}
\usage{
estimate_ts(species, TL, units = "mm")
}
\arguments{
\item{species}{A vector containing the species' scientific name (Clupea
pallasii, Engraulis mordax, Sardinops sagax, Scomber japonicus, or
Trachurus symmetricus).}

\item{TL}{A vector containing total length.}

\item{units}{Length units (e.g., "mm" or "cm").}
}
\value{
A data frame with target strength, backscattering coefficient, and
weight estimates.
}
\description{
Estimate target strength of CPS from total length
}
\examples{
estimate_ts("Sardinops sagax", TL, units = "mm")
}
