% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological.R
\name{estimate_weight}
\alias{estimate_weight}
\title{Estimate CPS weight}
\usage{
estimate_weight(scientificName, totalLength_mm, model.type = "OLS", season)
}
\arguments{
\item{scientificName}{A vector containing the species' scientific name (Clupea
pallasii, Engraulis mordax, Sardinops sagax, Scomber japonicus, or
Trachurus symmetricus).}

\item{totalLength_mm}{A vector containing total length (TL), in millimeters.}

\item{model.type}{The model used to estimate weight, either \code{OLS} (Ordinary Least Squares) or \code{GLM} (Generalized Linear Model).}

\item{season}{The season in which the specimen was collected, either \code{spring} or \code{summer}.}
}
\value{
Specimen weight, in grams.
}
\description{
Calculate the estimated weight of CPS species from a known length measurement, using the mass-length relationship
models presented in Palance et al. 2019.
}
\examples{
estimate_weight("Sardinops sagax", 300, model.type = "OLS", "summer")
}
