% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.R
\name{extract_cal_ecs}
\alias{extract_cal_ecs}
\title{Extract calibration results from Echoview calibration supplement (.ecs) files.}
\usage{
extract_cal_ecs(filename)
}
\arguments{
\item{filename}{An string containing the name of the calibration file.}
}
\value{
A data frame containing calibration results.
}
\description{
Extract calibration results from Echoview calibration supplement (.ecs) files.
}
