% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cal.R
\name{extract_cal_fm}
\alias{extract_cal_fm}
\title{Extract calibration results from calibrations in FM-mode.}
\usage{
extract_cal_fm(
  filename,
  vessel.name = NA_character_,
  survey.name = NA_character_,
  cal.group = NA_character_
)
}
\arguments{
\item{filename}{An string containing the name of the calibration file.}

\item{vessel.name}{A string containing the vessel name.}

\item{survey.name}{A string containing the survey name.}

\item{cal.group}{A string containing the name of the calibration group.}
}
\value{
A data frame containing frequency, gain, beam angles, and beam offsets.
}
\description{
Extract calibration results from ER60 (.txt) or EK80 (.xml) software.
}
