% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{extract_ctd}
\alias{extract_ctd}
\title{Extract header and cast information from CTD and UCTD casts.}
\usage{
extract_ctd(header.filename, cast.filename, type)
}
\arguments{
\item{header.filename}{Name of header file.}

\item{cast.filename}{Name of cast file.}

\item{type}{Cast type (CTD or UCTD).}
}
\description{
Extract header and cast information from CTD and UCTD casts.
}
