% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{extract_ctd_header}
\alias{extract_ctd_header}
\title{Extract header information from CTD and UCTD casts.}
\usage{
extract_ctd_header(header.filename, type)
}
\arguments{
\item{header.filename}{Name of header file.}

\item{type}{Cast type (CTD or UCTD).}
}
\value{
A data frame containing header information.
}
\description{
Extract header information from CTD and UCTD casts.
}
