% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_gps}
\alias{extract_gps}
\title{Create gps.csv files used by Echoview.}
\usage{
extract_gps(df, lat = df$lat, long = df$long, date_time = df$datetime)
}
\arguments{
\item{lat}{Latitude in decimal degrees.}

\item{long}{Longitude in decimal degrees.}

\item{date_time}{Date/time field in ISO 8601 date format (e.g., \%Y-\%m-\%d \%H:\%M:\%S).}
}
\value{
A data frame containing  \code{Date}, \code{Time}, \code{latitude}, and \code{longitude}.
}
\description{
Create gps.csv files used by Echoview.
}
