% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{fancy_sci}
\alias{fancy_sci}
\title{Format ggplot2 axis labels to fancy scientific.}
\usage{
fancy_sci(l)
}
\arguments{
\item{l}{A simple feature object.}
}
\value{
A text string as an expression.
}
\description{
Format ggplot2 axis labels to fancy scientific.
}
\examples{
p <- ggplot(mtcars, aes(wt, mpg))
p + geom_point() + scale_y_continuous(labels = fancy_scientific)
}
