% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{find_hull}
\alias{find_hull}
\title{Draw convex hull polygon around a collection of points}
\usage{
find_hull(df)
}
\arguments{
\item{df}{A data frame containing the latitude (\code{lat}) and longitude
(\code{long}) of point features.}
}
\value{
A data frame with convex hull vertices.
}
\description{
Draw convex hull polygon around a collection of points
}
\examples{
find_hull(df)
}
