% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTMB.R
\name{fitTMB}
\alias{fitTMB}
\title{Fit ATM model in TMB}
\usage{
fitTMB(
  X_guyk,
  coords_gz,
  satellite_iz = NULL,
  conventional_hz = NULL,
  survey_jz = NULL,
  E_guy = NULL,
  uy_tz = NULL,
  duration_u = NULL,
  cpp_version = FishStatsUtils::get_latest_version(package = "ATM"),
  tmb_dir = system.file("executables", package = "ATM"),
  run_dir = getwd(),
  compile_dir = run_dir,
  log2steps = 20,
  sigma2 = 4^2,
  spde_aniso = NULL,
  use_REML = TRUE,
  build_model = TRUE,
  run_model = TRUE,
  start_param_list = NULL,
  alpha_ratio_bounds = 1,
  diffusion_bounds = 0,
  constant_tail_probability = 1e-08,
  ...
)
}
\value{
Object of class \code{fitTMB}, containing parameter estimates and predicted time-integrated movement
}
\description{
\code{fitTMB} fits the Advection-taxis model in TMB
}
\author{
James Thorson
}
