% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapping.R
\name{get_basemap}
\alias{get_basemap}
\title{Create basemap}
\usage{
get_basemap(
  nav,
  states,
  countries,
  labels,
  bathy,
  boundaries = NULL,
  add.labels = TRUE,
  crs
)
}
\arguments{
\item{nav}{Data frame containing navigation data, used to set map boundaries.}

\item{states}{Simple feature with state boundaries.}

\item{countries}{Simple feature with country boundaries.}

\item{labels}{A data frame with names and locations of landmark labels.}

\item{bathy}{Simple feature with bathymetr contours.}

\item{boundaries}{User-specified bounding box (optional).}

\item{add.labels}{Add landmark labels (TRUE/FALSE).}

\item{crs}{Coordinate reference system (CRS) for map projection.}
}
\value{
A ggplot2 map object.
}
\description{
Create basemap
}
\examples{
get_basemap(nav, states, countries, landmarks, bathy, crs = 3310)
}
