% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATM_inference.R
\name{icd2phecode}
\alias{icd2phecode}
\title{Mapping the disease code from icd10 to phecode}
\usage{
icd2phecode(rec_data)
}
\arguments{
\item{rec_data}{input data which use ICD10 encoding; please refer to the internal example data \code{HES_icd10_example} for the formatting of the data.}
}
\value{
a data frame where most entries are mapped from ICD10 code to phecode
}
\description{
Mapping the disease code from icd10 to phecode; the mapping are based on https://phewascatalog.org/phecodes;
The input if using ICD-10 should be a string  f numbers and capital letters only. For example, "I25.1" should be "I251".
}
\examples{
phecode_data <- icd2phecode(HES_icd10_example)
}
