\name{interaction}
\alias{interaction}
\alias{interaction.combine}
\alias{interaction.predesign}
\title{Interaction Terms}
\description{Create interaction terms.}
\usage{
interaction (x, y=NULL, E, s, w=NULL)
interaction.combine (x, op)
interaction.predesign (x)
}
\arguments{
\item{x}{}
\item{y}{}
\item{E}{}
\item{s}{}
\item{w}{}
\item{op}{}
}


