% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.nan.df}
\alias{is.nan.df}
\title{Identify NaN values in a data frame.}
\usage{
\method{is.nan}{df}(df)
}
\arguments{
\item{df}{A data frame containing NaN values.}
}
\value{
The indices of NaN elements in a data frame.
}
\description{
Identify NaN values in a data frame.
}
\examples{
is.nan.df(df)
}
