% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{isNaNdf}
\alias{isNaNdf}
\title{Identify NaN values in a data frame.}
\usage{
isNaNdf(df)
}
\arguments{
\item{df}{A data frame that may contain NaN values.}
}
\value{
The indices of NaN elements in a data frame.
}
\description{
Identify NaN values in a data frame.
}
\examples{
isNaNdf(df)
}
