% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta.R
\name{K2C}
\alias{K2C}
\title{Convert Kelvin to degrees C}
\usage{
K2C(K)
}
\arguments{
\item{K}{numeric value of temperature in K.}
}
\value{
\code{K2C} returns the equivalent temperature in degrees C.
}
\description{
Convert Kelvin to degrees C
}
\author{
Dan Kelley
}
