\name{linear}
\alias{atm.ls}
\alias{evaluate.linear}
\alias{fit.linear}
\alias{linear}
\alias{linear.design}
\alias{linear.estimate}
\alias{linear.expand}
\alias{linear.ftest}
\alias{oospfitted.linear}
\alias{oospprint.linear.design}
\alias{oospprint.linear.estimate}
\alias{oospprint.summary.linear}
\alias{oospsummary.linear}
\alias{reset.linear}
\title{Linear Terms}
\description{Create linear terms}
\usage{
atm.ls (x, y, w)
\method{evaluate}{linear}(t, u, \dots)
\method{fit}{linear}(t, y, pump, \dots)
linear (x, y=NULL, key=x, E, s, w=NULL)
linear.design (key)
linear.estimate (labs, th=rep (0, length (labs) ) )
linear.expand (n, x)
linear.ftest (t, y, \dots)
\method{oospfitted}{linear}(t, \dots)
\method{oospprint}{linear.design}(d, \dots)
\method{oospprint}{linear.estimate}(e, \dots)
\method{oospprint}{summary.linear}(s, \dots)
\method{oospsummary}{linear}(t, y, \dots)
\method{reset}{linear}(t, \dots)
}
\arguments{
\item{x}{}
\item{y}{}
\item{w}{}
\item{t}{}
\item{u}{}
\item{pump}{}
\item{key}{}
\item{E}{}
\item{s}{}
\item{labs}{}
\item{th}{}
\item{n}{}
\item{d}{}
\item{e}{}
\item{\dots}{}
}


