% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATM_inference.R
\name{longdata2diseasematrix}
\alias{longdata2diseasematrix}
\title{Title}
\usage{
longdata2diseasematrix(rec_data)
}
\arguments{
\item{rec_data}{A diagnosis data frame with three columns; format data as HES_age_example; first column is individual ids (eid), second column is the disease code (diag_icd10);
third column is the age at diagnosis (age_diag). Note for each individual, we only keep the first onset of each diseases. Therefore, if there are multiple incidences of the same disease
within each individual, the rest will be ignored.}
}
\value{
a disease matrix with first column being the individual ids, columns follows are diseases with 0,1 coding.
}
\description{
Title
}
\examples{
disease_matrix <- longdata2diseasematrix(HES_age_example)
}
