% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{map_bounds}
\alias{map_bounds}
\title{Calculate map boundaries from lat/lon input.}
\usage{
map_bounds(lat, lon, pad = 0.05)
}
\arguments{
\item{lat}{Latitude in decimal degrees.}

\item{lon}{Longitude in decimal degrees.}

\item{pad}{Percentage that map boundaries are extended beyond the input data.}
}
\value{
A data frame containing the range of \code{lat} and \code{lon}.
}
\description{
Calculate map boundaries from lat/lon input.
}
