% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matexp.R
\name{matexp}
\alias{matexp}
\title{Matrix exponential and Euler approximation}
\usage{
matexp(mat_gg, log2steps = Inf)
}
\description{
\code{matexp} implements the matrix exponential and/or sparse Euler approximation
}
\author{
James Thorson
}
