% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{mbar2Pa}
\alias{mbar2Pa}
\title{Convert from millibars to Pascals}
\usage{
mbar2Pa(mbar)
}
\arguments{
\item{mbar}{numeric value of pressure in millibars.}
}
\value{
\code{mbar2Pa} returns pressure in Pa.
}
\description{
Convert from millibars to Pascals
}
