% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{mvp_export_template_ecs}
\alias{mvp_export_template_ecs}
\title{Export template ECS file for MVP200 casts}
\usage{
mvp_export_template_ecs(
  raw_descent_df,
  binned_df,
  cast_id,
  output_dir,
  template_path,
  tx_depth = 2
)
}
\arguments{
\item{raw_descent_df}{Data frame containing raw descent data.}

\item{binned_df}{Data frame containing binned descent data.}

\item{cast_id}{Data frame containing binned descent data.}

\item{output_dir}{Data frame containing binned descent data.}

\item{template_path}{Path to ECS template file.}

\item{tx_depth}{Depth of probe during transits (meters, default = 2).}
}
\value{
Exports a new ECS file in the specified directory.
}
\description{
Export template ECS file for MVP200 casts
}
