% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{mvp_parse_coordinate}
\alias{mvp_parse_coordinate}
\title{Extract coordinates from MVP200 cast file}
\usage{
mvp_parse_coordinate(coord_str)
}
\arguments{
\item{coord_str}{String containing lat/long value(s).}
}
\value{
Coordinate value(s) in decimal degrees.
}
\description{
Extract coordinates from MVP200 cast file
}
