% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ctd.R
\name{mvp_process_file}
\alias{mvp_process_file}
\title{Calculate depth-integrated average values}
\usage{
mvp_process_file(file_path, sampling_rate_hz = 25)
}
\arguments{
\item{file_path}{Path to MVP200 cast file.}

\item{sampling_rate_hz}{Sampling rate (Hz, default = 25.0).}
}
\value{
A list containing the downcast values, binned values, and metadata.
}
\description{
Calculate depth-integrated average values
}
