% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{neighbor_quantifier}
\alias{neighbor_quantifier}
\title{Calculate statistics around points}
\usage{
neighbor_quantifier(
  small.file,
  large.file,
  variable = NULL,
  method = "sp",
  crs = 4326,
  radius = 20,
  length = 100,
  plot.fig = TRUE,
  save.fig = TRUE,
  fig.name = NULL
)
}
\arguments{
\item{small.file}{A data frame containing points around which statistics are computed; must contain 'lat' and 'long'.}

\item{large.file}{A data frame containing points upon which statistics are computed; must contain 'lat' and 'long'.}

\item{variable}{Column name upon which to compute statistics}

\item{method}{Either 'sp' to compute spatial statistics in Cartesian space, using the {sp} package, or 'sf' to compute
statistics in geographic space using the {sf} package.}

\item{crs}{Coordinate reference system for using the 'sf' method (default is 4326 for WGS84 projection)}

\item{radius}{Length of radius (in nautical miles) around points in \code{small.file} used to compute spatial polygons}

\item{length}{Length of vector used to create polygon nodes}

\item{plot.fig}{Plot results (T/F)}

\item{save.fig}{Save figure (T/F)}

\item{fig.name}{Name of figure.}
}
\value{
A data frame with the number (n), mean, and standard deviation of the \code{variable} within each polygon.
}
\description{
Calculate statistics around points
}
