% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{Pa2mbar}
\alias{Pa2mbar}
\title{Convert from Pascls to millibars}
\usage{
Pa2mbar(Pa)
}
\arguments{
\item{Pa}{numeric value of pressure in Pascals.}
}
\value{
\code{Pa2mbar} returns pressure in millibars.
}
\description{
Convert from Pascls to millibars
}
