% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{phecode_icd10}
\alias{phecode_icd10}
\title{ICD-10 ↔ PheCode mapping}
\format{
A data frame/tibble. Common columns include:
\describe{
\item{ICD10}{ICD-10 code (character).}
\item{PheCode}{PheCode (character).}
\item{Excl..Phecodes}{ancestor PheCode range (character).}
\item{Excl..Phenotypes}{ancestor PheCode name (character).}
}
}
\usage{
phecode_icd10
}
\description{
Mapping table between ICD-10 codes and PheCodes.
}
\examples{
head(phecode_icd10)
}
\keyword{datasets}
