% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{phecode_icd10cm}
\alias{phecode_icd10cm}
\title{ICD-10-CM <-> PheCode mapping}
\format{
A data frame/tibble. Common columns include:
\describe{
\item{ICD10}{ICD-10-CM code (character).}
\item{phecode}{PheCode (character).}
\item{exclude_range}{ancestor PheCode range (character).}
\item{exclude_name}{ancestor PheCode name (character).}
}
}
\usage{
phecode_icd10cm
}
\description{
Mapping table between ICD-10-CM codes and PheCodes.
}
\examples{
head(phecode_icd10cm)
}
\keyword{datasets}
