% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sounding_plot.R
\name{plot}
\alias{plot}
\title{Plot atmospheric sounding data}
\arguments{
\item{sounding}{the result of a call to \code{\link[=readAtmosphericSounding]{readAtmosphericSounding()}}.}

\item{item}{character value indicating what to plot. At the moment,
the only choices is the default, \code{"DWPT+TEMP"}.}

\item{mar}{4-element numeric vector specifying margins, in the standard way
(i.e. this is handed to \code{\link[=par]{par()}} by this function). The default
is set to give space for height indications in the right margin.}

\item{mgp}{3-element numeric vector specifying axis spacing, in the usual way
(i.e. this is handed to \code{\link[=par]{par()}} by this function). The default removes some
whitespace in the margins.}
}
\description{
This is a preliminary version.  Over time, it might be evolved to
plot something nearer what is shown at Reference 1.
}
\examples{
library(atmosphere)
# Normally, would download the file with downloadAtmosphericSounding()
file <- system.file("extdata", "Shearwater_2024-07-11.txt.gz", package = "atmosphere")
sounding <- readAtmosphericSounding(file)
plot(sounding)

}
\references{
\enumerate{
\item \url{https://www.noaa.gov/jetstream/upperair/skew-t-log-p-diagrams}
\item Visual Explanation of Meteorological Skew T Log P Sounding Diagrams, 2019.
\url{https://www.youtube.com/watch?v=7p7c85hhgOo}.
}
}
\author{
Dan Kelley
}
