% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_partial_dependence.R
\name{plot_partial_dependence}
\alias{plot_partial_dependence}
\title{Make partial dependence plot including confidence interval}
\usage{
plot_partial_dependence(
  x,
  train,
  filename = paste0(getwd(), "/pdp"),
  n_samples = 0,
  width,
  height,
  probs = c(0.025, 0.975),
  ...
)
}
\description{
\code{plot_partial_dependence} makes a PDP plot including confidence interval from fixed-effect covariance
}
\author{
James Thorson
}
