% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sounding_plot.R
\name{plotAtmosphericSounding}
\alias{plotAtmosphericSounding}
\title{Plot atmospheric sounding data}
\usage{
plotAtmosphericSounding(sounding, item = "DWPT+TEMP")
}
\arguments{
\item{sounding}{the result of a call to \code{\link[=readAtmosphericSounding]{readAtmosphericSounding()}}.}

\item{item}{character value indicating what to plot. At the moment,
the only choices is the default, \code{"DWPT+TEMP"}.}
}
\description{
Plot atmospheric sounding data
}
\examples{
library(atmosphere)
soundingFile <- downloadAtmosphericSounding(stationName = "Shearwater")
sounding <- readAtmosphericSounding(soundingFile)
par(mar = c(3, 3, 1, 3), mgp = c(2, 0.7, 0))
plotAtmosphericSounding(sounding)
unlink(soundingFile)

}
\author{
Dan Kelley
}
