\name{plots}
\alias{oosplines.categorical}
\alias{oosplines.term}
\alias{oospplot.atm}
\alias{oospplot.categorical}
\alias{oospplot.interaction}
\alias{oospplot.term}
\title{Plot Functions}
\description{Plot functions.}
\usage{
\method{oosplines}{categorical}(t, \dots)
\method{oosplines}{term}(t, \dots)
\method{oospplot}{atm}(m, which=NA, \dots)
\method{oospplot}{categorical}(t, y=NULL, s, main=term.label (t),
	lwd=2, col=rgb (0, 0.6, 0.1), \dots )
\method{oospplot}{interaction}(t, \dots)
\method{oospplot}{term}(t, y=NULL, s, newplot=TRUE, main=term.label (t),
	lwd=2, col=rgb (0, 0.6, 0.1), \dots)
}
\arguments{
\item{t}{}
\item{m}{}
\item{which}{}
\item{y}{}
\item{s}{}
\item{main}{}
\item{lwd}{}
\item{col}{}
\item{newplot}{}
\item{\dots}{}
}



