% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitR.R
\name{print.fitR}
\alias{print.fitR}
\title{Print parameter estimates}
\usage{
\method{print}{fitR}(x, ...)
}
\arguments{
\item{x}{Output from \code{\link{fitR}}}

\item{...}{Not used}
}
\description{
Print parameter estimates and standard errors.
}
