% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTMB.R
\name{print.fitTMB}
\alias{print.fitTMB}
\title{Print parameter estimates}
\usage{
\method{print}{fitTMB}(x, ...)
}
\arguments{
\item{x}{Output from \code{\link{fitTMB}}}

\item{...}{Not used}
}
\description{
Print parameter estimates and standard errors.
}
