% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{project_sf}
\alias{project_sf}
\title{Convert simple feature to data frame with a different coordinate reference
system (CRS).}
\usage{
project_sf(sf, crs)
}
\arguments{
\item{sf}{A simple feature object.}

\item{crs}{The new coordinate reference system (CRS).}
}
\value{
A data frame with projected coordinates \code{X} and \code{Y}.
}
\description{
Convert simple feature to data frame with a different coordinate reference
system (CRS).
}
\examples{
project_sf(sf_obj, crs = 4326)
}
