% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sounding_read.R
\name{readSounding}
\alias{readSounding}
\title{Read atmospheric sounding data}
\usage{
readSounding(file)
}
\arguments{
\item{file}{character value indicating the name of a sounding file.
This file must have been downloaded with \code{\link[=downloadSounding]{downloadSounding()}}.}
}
\description{
Read an HTML file downloaded by \code{\link[=downloadSounding]{downloadSounding()}}.  This
code works by recognizing portions of the file as preformatted blocks, which
is a brittle method that will fail if the HTML format used on the server website
changes.  Obviously, it would be better to work with csv or json files, but
we have no way to know how to download such files, since the server provides
no documentation.
}
\examples{
library(atm)
file <- system.file("extdata", "Shearwater_2024-07-11.txt.gz", package = "atmosphere")
sounding <- readSounding(file)

}
\author{
Dan Kelley
}
