\name{sample}
\docType{data}
\alias{sample}
\title{Simulated Dataset}
\description{Simulated dataset for atm models.}
\usage{data (sample)}
\format{
A data frame representing a simulated dataset. Contains 200 realisations of 5 variables (1 response, 4 explanatories).
\tabular{rll}{
[,1] \tab x1 \tab numeric \cr
[,2] \tab x2  \tab numeric \cr
[,3] \tab x3 \tab factor \cr
[,4] \tab x4 \tab numeric \cr
[,5] \tab y  \tab numeric
}}
\keyword{datasets}



