% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{scs2posix}
\alias{scs2posix}
\title{Convert date and time from SCS format to POSIXct}
\usage{
scs2posix(date, time)
}
\arguments{
\item{date}{Date in SCS format.}

\item{time}{Time in SCS format.}
}
\value{
A date/time object in POSIXct format.
}
\description{
Convert date and time from SCS format to POSIXct
}
