% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{short_icd10}
\alias{short_icd10}
\title{Short labels (at most first for letters/digits) for ICD-10 codes}
\format{
A data frame/tibble. Common columns include:
\describe{
\item{ICD10}{ICD-10 code (character).}
\item{parent_phecode}{phecode of parent node (character).}
\item{Excl..Phecodes}{ancestor PheCode range (character).}
\item{Excl..Phenotypes}{ancestor PheCode name (character).}
\item{occ}{number of distinct patient in UKB}
}
}
\usage{
short_icd10
}
\description{
A lookup table mapping ICD-10 codes to concise human-readable labels.
}
\examples{
head(short_icd10)
}
\keyword{datasets}
