% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{short_icd10cm}
\alias{short_icd10cm}
\title{Short labels (at most first for letters/digits) for ICD-10-CM codes}
\format{
A data frame/tibble. Common columns include:
\describe{
\item{ICD10}{ICD-10 code (character).}
\item{parent_phecode}{phecode of parent node (character).}
\item{exclude_range}{ancestor PheCode range (character).}
\item{exclude_name}{ancestor PheCode name (character).}
\item{occ}{number of distinct patient in UKB}
}
}
\usage{
short_icd10cm
}
\description{
A lookup table mapping ICD-10-CM codes to concise human-readable labels.
}
\examples{
head(short_icd10cm)
}
\keyword{datasets}
