% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTMB.R
\name{simulate.fitTMB}
\alias{simulate.fitTMB}
\title{Print parameter estimates}
\usage{
\method{simulate}{fitTMB}(x, random_seed = NULL, ...)
}
\arguments{
\item{x}{Output from \code{\link{fitTMB}}}

\item{...}{Not used}
}
\description{
Predict habitat preference
}
