% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/skew.R
\name{skew}
\alias{skew}
\title{Determine skew skew-T diagrams}
\usage{
skew(pressure)
}
\arguments{
\item{pressure}{numeric vector holding the pressures in the
sounding dataset.}
}
\value{
\code{skew} returns the value to be added to temperature, as a
function of the supplied pressure.
}
\description{
Based on the plotted data range and figure geometry, this determines a
temperature skew that be equivalent to a 45 degree rotation of isotherms.  A
warning will be issued if the graph is far from square or the axes
cover unexpected ranges.
}
\author{
Dan Kelley
}
