% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{SNOMED_ICD10CM}
\alias{SNOMED_ICD10CM}
\title{SNOMED ↔ ICD-10(-CM) mapping (excerpt)}
\format{
A data frame/tibble. Common columns include:
\describe{
\item{SNOMED}{SNOMED CT concept identifier (character).}
\item{ICD10}{ICD-10 code (character), and/or}
\item{ICD10_name}{ICD-10-CM code (character).}
\item{SNOMED_description}{SNOMED readable explanation}
\item{occ}{ICD10 occurence in UKB}
}
}
\usage{
SNOMED_ICD10CM
}
\description{
A small mapping table used by functions such as \code{\link{icd2phecode}}
}
\examples{
head(SNOMED_ICD10CM)
}
\keyword{datasets}
