% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sounding_station.R
\name{stationNumber}
\alias{stationNumber}
\title{Find station number from name}
\usage{
stationNumber(name = "Shearwater")
}
\arguments{
\item{name}{character value indicating the name of the station.
So far, there are only two choices: \verb{"Shearwater"`` and }"Nashville"`.}
}
\description{
Find station number from name
}
\author{
Dan Kelley
}
