% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTMB.R
\name{summary.fitTMB}
\alias{summary.fitTMB}
\title{Check residuals etc.}
\usage{
\method{summary}{fitTMB}(x, what = "survey_residuals", n_samples = 250, working_dir = NULL, ...)
}
\description{
\code{summary.fitTMB} extracts commonly used quantities derived from a fitted VAST model
}
\details{
\code{summary.fitTMB} faciliates common queries for model output including:
}
