% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermodynamics.R
\name{TC2TK}
\alias{TC2TK}
\title{Convert degrees C to Kelvin}
\usage{
TC2TK(TC)
}
\arguments{
\item{TC}{numeric value of temperature in degrees C.}
}
\value{
\code{TC2TK} returns the equivalent temperature in Kelvin.
}
\description{
Convert degrees C to Kelvin
}
\author{
Dan Kelley
}
