% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta.R
\name{temperatureFromTheta}
\alias{temperatureFromTheta}
\title{Compute air in-situ temperature from potential temperature}
\usage{
temperatureFromTheta(theta, pressure, R = 287.052874, Cp = 1003.5)
}
\arguments{
\item{theta}{numeric value indicating potential temperature in
Kelvin.}

\item{pressure}{numeric value indicating pressure in Pa.}

\item{R}{dry-air gas constant, with the default value of
287.052874 from Reference 2.}

\item{Cp}{dry-air mass-specific heat, with the default value of
1003.5 J/kg/K from Reference 3.  (Note that 1005 is given in some
websites.)}
}
\value{
\code{temperatureFromTheta} returns in-situ temperature, in Kelvin.
}
\description{
In-situ temperature is computed as in Reference 1, by multiplying the
potential temperature by the ratio of actual to reference pressure,
raised to the power R/Cp.  Although both \code{R} and \code{Cp} can be specified in
this function, the default values are recommended for consistency.
}
\references{
\enumerate{
\item \url{https://en.wikipedia.org/wiki/Potential_temperature}
\item \url{https://en.wikipedia.org/wiki/Gas_constant}
\item \url{https://en.wikipedia.org/wiki/Table_of_specific_heat_capacities}
}
}
\author{
Dan Kelley, based on reference 1.
}
