% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/textInBox.R
\name{textInBox}
\alias{textInBox}
\title{Draw a box with text on top}
\usage{
textInBox(x, y, labels, font = 1, cex = 1, col = 1, bg = "white")
}
\arguments{
\item{x, y, labels, font, cex, col}{as for \code{\link[=text]{text()}}.}

\item{bg}{colour for the box underlay.}
}
\description{
This is useful for adding text to busy plots, as the altitude labels in
skew-T diagrams created with \code{\link[=plot.sounding]{plot.sounding()}}.
}
\examples{
plot(c(1, 2), c(10, 20), bg = "gray", type = "n")
rect(xleft = 0, xright = 3, ybottom = 5, ytop = 25, col = "gray")
abline(h = 15)
abline(v = 1.5)
textInBox(1.5, 15, "12345 m")

}
\author{
Dan Kelley
}
