% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/conversions.R
\name{TK2TC}
\alias{TK2TC}
\title{Convert Kelvin to degrees C}
\usage{
TK2TC(TK)
}
\arguments{
\item{TK}{numeric value of temperature in K.}
}
\value{
\code{TK2TC} returns the equivalent temperature in degrees C.
}
\description{
Convert Kelvin to degrees C
}
\author{
Dan Kelley
}
