% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{UKB_349_disease}
\alias{UKB_349_disease}
\title{List of 349 UK Biobank diseases (example)}
\format{
A data frame/tibble containing disease identifiers/names. Columns include:
\describe{
\item{diag_icd10}{Phecode (character).}
\item{occ}{number of distinct patient in UKB}
} @examples
head(UKB_349_disease)
}
\usage{
UKB_349_disease
}
\description{
A character vector or table listing the set of disease phenotypes used
in examples/vignettes.
}
\keyword{datasets}
