% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{UKB_HES_10topics}
\alias{UKB_HES_10topics}
\title{Example topic model output (10 topics, UKB HES)}
\format{
An array for UKB topic loadings.
Dimention is age, disease, topics. the ordering of disease is the same as UKB_349_disease.
}
\usage{
UKB_HES_10topics
}
\description{
An illustrative result object/table from a 10-topic model fit to UKB HES-like data;
used for examples, plotting, and tests.
}
\examples{
head(UKB_HES_10topics)
}
\keyword{datasets}
