% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xtrct}
\alias{xtrct}
\title{Extract nodes from XML file.}
\usage{
xtrct(doc, target)
}
\arguments{
\item{doc}{An XML object.}

\item{target}{A regular expression.}
}
\value{
A text vector.
}
\description{
Extract nodes from XML file.
}
\seealso{
See \url{https://rpubs.com/hrbrmstr/xml2power} for original
functions; used by \code{\link{xtrct_df}}
}
