% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{xtrct_df}
\alias{xtrct_df}
\title{Extract child nodes and values from node list.}
\usage{
xtrct_df(doc, top)
}
\arguments{
\item{doc}{An XML object.}

\item{top}{A named node.}
}
\value{
A data frame of named child vectors.
}
\description{
Extract child nodes and values from node list.
}
\seealso{
See \url{https://rpubs.com/hrbrmstr/xml2power} for original
functions, and related helper function \code{\link{xtrct}}
}
