% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biological.R
\name{convert_length}
\alias{convert_length}
\title{Convert CPS length}
\usage{
convert_length(scientificName, L.in, from, to)
}
\arguments{
\item{scientificName}{A vector containing the species' scientific name (Clupea
pallasii, Engraulis mordax, Sardinops sagax, Scomber japonicus, or
Trachurus symmetricus).}

\item{L.in}{The length measurement to be converted (unitless) .}

\item{from}{The measurement type from which length is converted.}

\item{to}{The measurement type to which length is converted.}
}
\value{
The converted length (unitless).
}
\description{
Converts between various length measures (total length (TL), standard length (SL), and fork length (FL)) for CPS species,
using the major-axis regression models presented in Palance et al. 2019. Not all conversions are possible. In such cases, \code{NA} is returned.
}
\examples{
convert_length("Sardinops sagax", 20, model.type = "OLS", "summer")
}
