% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lapse_rate.R
\name{dryLapseRate}
\alias{dryLapseRate}
\title{Dry lapse rate in degC / m}
\usage{
dryLapseRate(g = 9.7986, C_P = 1004)
}
\arguments{
\item{g}{numeric value of the acceleration due to gravity at sea level,
with the default being as listed in Reference 1, converted to
J/(kg*K).}

\item{C_P}{numeric value of the specific heat of air at sea level,
with the default being as listed in Reference 1, converted to m/s^2.}
}
\value{
\code{dryLapseRate} returns the lapse rate, in degrees C (or Kelvin) per
metre.
}
\description{
Dry lapse rate in degC / m
}
\references{
\enumerate{
\item \url{https://pds-atmospheres.nmsu.edu/education_and_outreach/encyclopedia/adiabatic_lapse_rate.htm}
}
}
\author{
Dan Kelley, based on Reference 1.
}
