% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{estimate_bootstrap}
\alias{estimate_bootstrap}
\title{Calculate bootstrap estimates of species biomass}
\usage{
estimate_bootstrap(nasc.df, clf.df, stratum.num, stratum.area,
  species = NULL, do.lf = FALSE, boot.number = 0)
}
\arguments{
\item{nasc.df}{A data frame containing total backscatter from all CPS
(\code{cps.nasc}), acoustic proportions (\code{prop.spp}), and
weight-specific backscattering cross sections (\code{sigmawg.spp}) in a
given stratum}

\item{clf.df}{A data frame containing cluster length frequencies for a given
\code{species} (\code{cps.nasc}), acoustic proportions (\code{prop.spp}),
and weight-specific backscattering cross sections (\code{sigmawg.spp}) in a
given stratum}

\item{stratum.num}{A vector containing stratum number.}

\item{stratum.area}{A vector containing stratum area (in square meters).}

\item{species}{A vector containing the species' scientific name (Clupea
pallasii, Engraulis mordax, Sardinops sagax, Scomber japonicus, or
Trachurus symmetricus).}

\item{do.lf}{Return a length frequency (\code{TRUE/FALSE})}

\item{boot.number}{A vector defining the number of bootstrap samples.}
}
\value{
A data frame containing estimates of biomass and abundance or a list
of biomass, abundance, and length frequency.
}
\description{
Calculate bootstrap estimates of species biomass
}
\examples{
estimate_bootstrap(nasc.df, clf.df, stratum.num, stratum.area,
"Engraulis mordax", do.lf = TRUE, boot.num = 1000)
}
