% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{estimate_point}
\alias{estimate_point}
\title{Calculate point estimate of species biomass}
\usage{
estimate_point(df, stratum.area, species)
}
\arguments{
\item{df}{A data frame containing total backscatter from all CPS
(\code{cps.nasc}), acoustic proportions (\code{prop.spp}), and
weight-specific backscattering cross sections (\code{sigmawg.spp}) in a
given stratum}

\item{stratum.area}{A vector containing stratum areas (in square meters).}

\item{species}{A vector containing the species' scientific name (Clupea
pallasii, Engraulis mordax, Sardinops sagax, Scomber japonicus, or
Trachurus symmetricus).}
}
\value{
A data frame with the stratum areas and point estimates of total
biomass.
}
\description{
Calculate point estimate of species biomass
}
\examples{
estimate_point(df, stratum.area, "Engraulis mordax")
}
