% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/atm.R
\name{estimate_ts}
\alias{estimate_ts}
\title{Estimate target strength of CPS from total length}
\usage{
estimate_ts(species, TL, units = "mm")
}
\arguments{
\item{species}{A value or vector containing the species' scientific name (Clupea
pallasii, Engraulis mordax, Sardinops sagax, Scomber japonicus, or
Trachurus symmetricus).}

\item{TL}{A value or vector containing total length measurements.}

\item{units}{Length measurement units (e.g., "mm" or "cm"). If mm, lengths are converted to cm internally but returned in cm.}
}
\value{
A data frame with species name, total length, target strength, backscattering coefficients, and
weight estimates.
}
\description{
Estimate target strength of CPS from total length
}
\examples{
estimate_ts("Sardinops sagax", TL, units = "mm")
}
