% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{extract_csv}
\alias{extract_csv}
\title{Extract vertically integrated backscatter from CSV files}
\usage{
extract_csv(filename)
}
\arguments{
\item{filename}{Name or path to CSV file exported from Echoview.}
}
\value{
A data frame containing vertically integrated backscatter data.
}
\description{
Extract vertically integrated backscatter from CSV files
}
