% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATM_inference.R
\name{icd2phecode}
\alias{icd2phecode}
\title{Title mapping the disease code from icd10 to phecode; the mapping are based on https://phewascatalog.org/phecodes}
\usage{
icd2phecode(rec_data)
}
\arguments{
\item{rec_data}{input data which use ICD10 encoding; please refer to the internal example data \code{HES_icd10_example} for the formatting of the data.}
}
\value{
a data frame where most entries are mapped from ICD10 code to phecode
}
\description{
Title mapping the disease code from icd10 to phecode; the mapping are based on https://phewascatalog.org/phecodes
}
\examples{
phecode_data <- icd2phecode(HES_icd10_example)
}
