% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_data}
\alias{make_data}
\title{Build data input for ATM model}
\usage{
make_data(
  Cov_stars,
  formula_taxis,
  formula_diffusion,
  coords_gz,
  uy_tz = NULL,
  conventional_hz = NULL,
  satellite_iz = NULL,
  survey_jz = NULL,
  fishery_fz = NULL,
  E_guy = NULL,
  duration_u = NULL,
  cpp_version = "R",
  log2steps = 20,
  spde_aniso = NULL,
  alpha_ratio_bounds = 1,
  diffusion_bounds = 0,
  constant_tail_probability = 1e-08,
  movement_penalty = 0
)
}
\value{
Object of class \code{make_data}, containing inputs to function \code{\link{make_model}}
}
\description{
\code{make_data} builds a tagged list of data inputs used by TMB for running the model
}
\author{
James Thorson
}
