% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_data.R
\name{make_data}
\alias{make_data}
\title{Build data input for ATM model}
\usage{
make_data(
  X_guyk,
  loc_gz,
  uy_tz,
  satellite_iz,
  survey_jz,
  cpp_version = "R",
  duration_u
)
}
\value{
Object of class \code{make_data}, containing inputs to function \code{\link{make_model}}
}
\description{
\code{make_data} builds a tagged list of data inputs used by TMB for running the model
}
\author{
James Thorson
}
