% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fitTMB.R
\name{predict.fitTMB}
\alias{predict.fitTMB}
\title{Print parameter estimates}
\usage{
\method{predict}{fitTMB}(
  x,
  newdata,
  formula,
  varname = "alpha_k",
  origdata = NULL,
  prediction_type = 1,
  seed = NULL
)
}
\arguments{
\item{x}{Output from \code{\link{fitTMB}}}

\item{...}{Not used}
}
\description{
Predict habitat preference
}
