% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spatial.R
\name{project_df}
\alias{project_df}
\title{Project a data frame with geographic coordinates to a different coordinate
reference system (CRS).}
\usage{
project_df(df, lat, long, crs.geog = 4326, crs.proj)
}
\arguments{
\item{df}{A data frame with geographic coordinates (latitude and longitude,
in decimal degrees) and a WGS84 CRS (4326).}

\item{lat}{The quoted column name containing latitude values (decimal degrees).}

\item{long}{The quoted column name containing longitude values (decimal degrees).}

\item{crs.proj}{The new projected coordinate reference system.}
}
\value{
A data frame with a different CRS.
}
\description{
Project a data frame with geographic coordinates to a different coordinate
reference system (CRS).
}
\examples{
project_df(df, lat, long, crs.proj = 3310)
}
