% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sounding_read.R
\name{readAtmosphericSounding}
\alias{readAtmosphericSounding}
\title{Read atmospheric sounding data}
\usage{
readAtmosphericSounding(file)
}
\arguments{
\item{file}{character value indicating the name of a sounding file.
This file must have been downloaded with \code{\link[=downloadAtmosphericSounding]{downloadAtmosphericSounding()}}.}
}
\description{
Read atmospheric sounding data
}
\author{
Dan Kelley
}
