% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATM_simulations.R
\name{simulate_genetic_disease_from_topic}
\alias{simulate_genetic_disease_from_topic}
\title{Step 2 of the two step procedure to simulate Genetic-Disease-topic structure;
output
5 types of G-disease-topic structure
3 sets of SNPs.
\enumerate{
\item SNP -> disease -> topic: SNP id: 1-20; disease id: para$D+1; topic id: 1
\item SNP*Topic -> disease: SNP id: 41-60; disease id: para$D + 2; topic id: 1
\item SNP -> Topic -> disease; SNP -> disease: SNP id: 21-(20+varying number(v2t)); disease id: para$D + 3; topic id: 1
\item SNP -> Topic -> disease; SNP + SNP^2 -> disease: SNP id: 21-(20+varying number(v2t)); disease id: para$D + 4; topic id: 1
\item SNP -> Topic + Topic^2 -> disease; SNP -> disease: SNP id: 21-(20+varying number(v2t)); disease id: para$D + 5; topic id: 1
}}
\usage{
simulate_genetic_disease_from_topic(
  para,
  genetics_population,
  causal_disease,
  disease_number,
  ds_per_idv = 6.1,
  itr_effect = 0,
  topic2disease = 2,
  v2t = 20,
  liability_thre = 0.8
)
}
\arguments{
\item{para}{Simulated topic loadings; should be the first element of the output of simulate_topics.}

\item{genetics_population}{Simulated genotypes, should be the second element of the output of simulate_topics.}

\item{causal_disease}{Simulated causal diseases, should be the third element of the output of simulate_topics.}

\item{disease_number}{Number of diseases to simulate from the topic. The final number of disease will be disease_number + 5.}

\item{ds_per_idv}{Number of disease per individual, default 6.1, which is the UKB number.}

\item{itr_effect}{The simulated interaction effect, default set to 0.}

\item{topic2disease}{Simulated topic to disease effect, default set to 2.}

\item{v2t}{number of variant that have an effect on topic, has to be the same value as the previous simulate_topics function.}

\item{liability_thre}{Threshold of liability for simulating disease, which equal to proportion of individual set to be healthy, e.g. 0.8 means the
top 20\% of liability is set to have the diseaes; default 0.8.}
}
\value{
a list, the first element is the simulated disease data (the only element you actually need);
the second element is a data list ; third element is the disease list (binary) representing #2 structure;
fourth element is the disease list (binary) representing #3 structure.
}
\description{
Step 2 of the two step procedure to simulate Genetic-Disease-topic structure;
output
5 types of G-disease-topic structure
3 sets of SNPs.
\enumerate{
\item SNP -> disease -> topic: SNP id: 1-20; disease id: para$D+1; topic id: 1
\item SNP*Topic -> disease: SNP id: 41-60; disease id: para$D + 2; topic id: 1
\item SNP -> Topic -> disease; SNP -> disease: SNP id: 21-(20+varying number(v2t)); disease id: para$D + 3; topic id: 1
\item SNP -> Topic -> disease; SNP + SNP^2 -> disease: SNP id: 21-(20+varying number(v2t)); disease id: para$D + 4; topic id: 1
\item SNP -> Topic + Topic^2 -> disease; SNP -> disease: SNP id: 21-(20+varying number(v2t)); disease id: para$D + 5; topic id: 1
}
}
\examples{
rslts <- simulate_topics(topic_number = 2,
          pop_sz=1000, disease2topic  = 0.1, v2t = 20 )
para_sim <- rslts[[1]]
genetics_population <- rslts[[2]]
causal_disease <- rslts[[3]]
reslt_ds <- simulate_genetic_disease_from_topic(para_sim, genetics_population,causal_disease,
disease_number = 20, itr_effect = 1,
topic2disease = 2, v2t = 20)
rec_data <- reslt_ds[[1]]
}
