% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATM_simulations.R
\name{simulate_topics}
\alias{simulate_topics}
\title{Step 1 of the two step procedure to simulate Genetic-Disease-topic structure;
all genetic effect are on the first topic only
5 types of G-disease-topic structure
3 sets of SNPs.
\enumerate{
\item SNP -> disease -> topic: SNP id: 1-20; disease id: para$D+1; topic id: 1
\item SNP*Topic -> disease: SNP id: 41-60; disease id: para$D + 2; topic id: 1
\item SNP -> Topic -> disease; SNP -> disease: SNP id: 21-(20+varying number(v2t)); disease id: para$D + 3; topic id: 1
\item SNP -> Topic -> disease; SNP + SNP^2 -> disease: SNP id: 21-(20+varying number(v2t)); disease id: para$D + 4; topic id: 1
\item SNP -> Topic + Topic^2 -> disease; SNP -> disease: SNP id: 21-(20+varying number(v2t)); disease id: para$D + 5; topic id: 1
}}
\usage{
simulate_topics(
  topic_number,
  num_snp = 100,
  pop_sz = 10000,
  disease2topic = 0,
  v2t = 20,
  snp2t = 0.04,
  snp2d = 0.15,
  liability_thre = 0.8
)
}
\arguments{
\item{topic_number}{Number of topics to simulate.}

\item{num_snp}{Number of SNPs in the simulation, default is 100; no less than 60.}

\item{pop_sz}{Number of individual to simulate, default set to 10,000.}

\item{disease2topic}{Diseases to topic effect size, default set to 0;}

\item{v2t}{Number of variants that are effective on topic 1, should be set between 0 and 20; default 20.}

\item{snp2t}{Effect sizes of SNP to topic effect, default 0.04. (informed by real UKB data)}

\item{snp2d}{Effect sizes of SNP to disease effect, default 0.15.}

\item{liability_thre}{Threshold of liability for simulating disease, which equal to proportion of individual set to be healthy, e.g. 0.8 means the
top 20\% of liability is set to have the diseaes; default 0.8.}
}
\value{
a list with three objects, each of it should be input to the simulate_genetic_disease_from_topic function;
first is a para object, which saves the topic information; second one is the simulated genotypes;
third is one simulated disease (binary) which is causal to the topic loading of the first topic.
}
\description{
Step 1 of the two step procedure to simulate Genetic-Disease-topic structure;
all genetic effect are on the first topic only
5 types of G-disease-topic structure
3 sets of SNPs.
\enumerate{
\item SNP -> disease -> topic: SNP id: 1-20; disease id: para$D+1; topic id: 1
\item SNP*Topic -> disease: SNP id: 41-60; disease id: para$D + 2; topic id: 1
\item SNP -> Topic -> disease; SNP -> disease: SNP id: 21-(20+varying number(v2t)); disease id: para$D + 3; topic id: 1
\item SNP -> Topic -> disease; SNP + SNP^2 -> disease: SNP id: 21-(20+varying number(v2t)); disease id: para$D + 4; topic id: 1
\item SNP -> Topic + Topic^2 -> disease; SNP -> disease: SNP id: 21-(20+varying number(v2t)); disease id: para$D + 5; topic id: 1
}
}
\examples{
rslts <- simulate_topics(topic_number = 2, disease2topic  = disease2topic, v2t = cont_v2t )
para_sim <- rslts[[1]]
genetics_population <- rslts[[2]]
causal_disease <- rslts[[3]]
reslt_ds <- simulate_genetic_disease_from_topic(para_sim, genetics_population,causal_disease,
disease_number = disease_number, itr_effect = itr_effect,
topic2disease = topic2disease, v2t = cont_v2t)
rec_data <- reslt_ds[[1]]
}
