% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sounding_station.R
\name{stationNumber}
\alias{stationNumber}
\title{Find station number from name}
\usage{
stationNumber(name = NULL)
}
\arguments{
\item{name}{character value indicating the name of the station.
Use \code{name=NULL} (the default) to get a list of known names.
This list is \emph{very} short, and not necessarily up-to-date!}
}
\description{
Find station number from name
}
\author{
Dan Kelley
}
