% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{summary}
\alias{summary}
\title{Summarize a sounding}
\arguments{
\item{x}{a \code{sounding} object, created with \code{\link[=readAtmosphericSounding]{readAtmosphericSounding()}}.}
}
\description{
Summarize a sounding
}
\examples{
library(atmosphere)
file <- system.file("extdata", "Shearwater_2024-07-11.txt.gz", package = "atmosphere")
sounding <- readAtmosphericSounding(file)
summary(sounding)

}
\author{
Dan Kelley
}
