% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theta.R
\name{T2theta}
\alias{T2theta}
\title{Compute air potential temperature from in-situ temperature}
\usage{
T2theta(temperature, pressure, R = 287.052874, Cp = 1003.5)
}
\arguments{
\item{temperature}{numeric value indicating in-situ temperature in
degrees C.}

\item{pressure}{numeric value indicating pressure in mbar (equivalently,
in hPa).}

\item{R}{dry-air gas constant, with the default value of
287.052874 from Reference 2.}

\item{Cp}{dry-air mass-specific heat, with the default value of
1003.5 J/kg/K from Reference 3.  (Note that 1005 is given in some
websites.)}
}
\value{
\code{T2theta} returns potential temperature, in degrees Celcius.
}
\description{
Potential temperature is computed as in Reference 1, by multiplying the
in-situ temperature by the ratio reference pressure to the actual pressure,
raised to the power R/Cp.  Although both \code{R} and \code{Cp} can be specified in
this function, the default values are recommended for consistency.
}
\references{
\enumerate{
\item \url{https://en.wikipedia.org/wiki/Potential_temperature}
\item \url{https://en.wikipedia.org/wiki/Gas_constant}
\item \url{https://en.wikipedia.org/wiki/Table_of_specific_heat_capacities}
}
}
\author{
Dan Kelley, based on reference 1.
}
