% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thermodynamics.R
\name{theta2T}
\alias{theta2T}
\title{Compute air in-situ temperature from potential temperature}
\usage{
theta2T(theta, pressure, Rd = 287.04, Cp = 1005.7)
}
\arguments{
\item{theta}{numeric value indicating potential temperature in
degrees C.}

\item{pressure}{numeric value indicating pressure in mbar (equivalently,
in hPa).}

\item{Rd, Cp}{dry-air gas constant and mass-specific heat capacity,
with the default values as in Reference 2.}
}
\value{
\code{theta2T} returns in-situ temperature, in degrees Celcius.
}
\description{
In-situ temperature is computed as in Chapter 3 of Reference 1, i.e. by
multiplying the potential temperature by the ratio of actual pressure to reference
pressure, raised to the power \code{Rd}/\code{Cp}. (The computation is done in Absolute
Temperature, with conversions to and from Celcius being done with \code{\link[=TK2TC]{TK2TC()}}
and \code{\link[=TC2TK]{TC2TK()}}, respectively.)  See also \code{\link[=T2theta]{T2theta()}}, which is the inverse
of \code{theta2T()}.
}
\references{
\enumerate{
\item Wallace, John M., and Peter Victor Hobbs. Atmospheric Science: An
Introductory Survey. 2nd ed. International Geophysics Series, volume 92.
Amsterdam; Boston: Elsevier Academic Press, 2006.
\item \url{https://github.com/cwebster2/pyMeteo}
}
}
\author{
Dan Kelley, with the default values for \code{Rd} and \code{cp}
taken from Reference 1.
}
