% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wind_barbs.R
\name{windBarbs}
\alias{windBarbs}
\title{Draw wind barbs}
\usage{
windBarbs(
  x,
  y,
  u,
  v,
  scale = 1,
  length = 0.5,
  step = 10,
  lwd = 1,
  col = 1,
  debug = FALSE
)
}
\arguments{
\item{x, y}{numeric values indicating the locations at which to
draw starting points of barbs.}

\item{u, v}{numeric values indicating the velocity components in
the x and y directions.}

\item{scale}{numeric value indicating the length of barbs,
as measured in the axis scale.}

\item{length}{numeric value indicating the ratio of flag length
to barb length.}

\item{step}{numeric value indicating the step between speed categories.
The default value of 10 is somewhat standard for speeds in knots.}

\item{lwd}{line width of barbs.}

\item{col}{colour of barbs.}

\item{debug}{integer that controls tell degree of debugging.  If this is
zero, work proceeds silently.  Otherwise, some debugging information will
be printed.}
}
\description{
Draw wind barbs on an existing plot.  This (at the moment) requires
that the plot aspect ratio be 1.
}
\examples{
library(atm)
n <- 10
L <- 10
x <- seq(-L, L, length.out = n)
y <- seq(-L, L, length.out = n)
plot(x, y, type = "n", xlab = "x", ylab = "y", asp = 1)
g <- expand.grid(x = x, y = y)
R <- sqrt(g$x^2 + g$y^2)
U <- 30 * R * exp(-R / (L / 2))
theta <- atan2(g$y, g$x)
g$u <- U * sin(theta)
g$v <- -U * cos(theta)
windBarbs(g$x, g$y, g$u, g$v, length=0.5, scale = 2)

}
\author{
Dan Kelley
}
