% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ATM_inference.R
\name{wrapper_ATM}
\alias{wrapper_ATM}
\title{Run ATM on diagnosis data.}
\usage{
wrapper_ATM(
  rec_data,
  topic_num,
  degree_free_num = 3,
  CVB_num = 5,
  save_data = F
)
}
\arguments{
\item{rec_data}{A diagnosis data frame with three columns; format data as HES_age_example; first column is individual ids, second column is the disease code;
third column is the age at diagnosis. Note for each individual, we only keep the first onset of each diseases. Therefore, if there are multiple incidences of the same disease
within each individual, the rest will be ignored.}

\item{topic_num}{Number of topics to infer.}

\item{degree_free_num}{control the parametric for of topic loadings: Degrees of freedom (d.f.) from 2 to 7 represent linear, quadratic polynomial, cubic polynomial, spline with one knot, spline with two knots, and spline with three knots. Default is set to 3.}

\item{CVB_num}{Number of runs with random initialization. The final output will be the run with highest ELBO value.}

\item{save_data}{A flag which determine whether full model data will be saved. If TRUE, a Results/ folder will be created and full model data will be saved. Default is set to be FALSE.}
}
\value{
Reture a list object with topic_loadings (of the best run), topic_weights (of the best run), ELBO_convergence (ELBO until convergence),
patient_list (list of eid which correspond to rows of topic_weights), ds_list (gives the ordering of diseases in the topic_loadings object), disease_number (number of total diseases), patient_number(total number of patients), topic_number (total number of topic),
topic_configuration (control the parametric for of topic loadings: Degrees of freedom (d.f.) from 2 to 7 represent linear, quadratic polynomial,
cubic polynomial, spline with one knot, spline with two knots, and spline with three knots. Default is set to 3.), multiple_run_ELBO_compare (ELBO of each runs).
}
\description{
Run ATM on diagnosis data to infer topic loadings and topic weights. Note one run of ATM on 100K individuals would take ~30min (defualt is 5 runs and pick the best fit);
if the data set is small and the goal is to infer patient-level topic weights (i.e. assign comorbidity profiles to individuals based on the disedases),
please use loading2weights.
}
\examples{
  HES_age_small_sample <- HES_age_example \%>\%
dplyr::slice_sample(prop = 0.1)
inference_results <- wrapper_ATM(HES_age_small_sample, topic_num = 10, CVB_num = 1)
}
